breaking: Use two modules to handle cjs / esmodule export differences
Because ProtractorFlake used a default export, and that is broken by the
es-module `default`, I am choosing to export _two_ entrypoints. One, the
`main` preserves compatibility with node's `require`, the other
`protractor-flake/es` is a es-modules/typescript entrypoint. In the
future I may consider a breaking cut over that promotes the es module
style to `main`.