breaking: Use Typescript (#96)
The introduction of types and the TS ecosystem (and the nice parity with protractor core being in TS) makes this a win for the project. I also feel better about the stability of TS over babel even if we miss out on immediate access to es-next features.

Although this technically won't break existing projects I am marking this as "breaking" just in case (semver gods forgive me).